+++ src/components/organisms/ImageItem.tsx
@@ -1,8 +1,7 @@
 import React, { useCallback, memo, ChangeEvent, useState } from "react";
 import { useParks } from "../../hooks/useParks";
+import { useImages } from "../../hooks/useImages";
 import { ImageItemDialog } from "../molecules/ImageItemDialog";
-import format from "date-fns/format";
-import ja from "date-fns/locale/ja";
 import {
   Card,
   CardContent,
@@ -55,6 +54,7 @@ export const ImageItem = memo((props: Props) => {
     isFollowed,
   } = props;
   const { parks } = useParks();
+  const { createdTime } = useImages();
   // checkboxの切り替え
   const handleCheckBoxChange = useCallback(
     (e: ChangeEvent<HTMLInputElement>) => {
@@ -66,16 +66,6 @@ export const ImageItem = memo((props: Props) => {
     },
     [checked, handleSelect, handleRemove]
   );
-
-  // 撮影日時をフォーマットする
-  const createdTime = (image: Image) => {
-    if (image.taken_at) {
-      const date = new Date(image.taken_at);
-      const formattedDate = format(date, "yyyy/M/d/(E)", { locale: ja });
-      return formattedDate;
-    }
-    return null;
-  };
 
   // DialogのOpen/Close
   const [imageOpen, setImageOpen] = useState<boolean>(false);
